requirements.txt
aiogram>=3.4.0
google-generativeai>=0.4.0
python-dotenv>=1.0.1
SQLAlchemy>=2.0.29
aiosqlite>=0.20.0
asyncpg>=0.29.0
APScheduler>=3.10.4
pydantic>=2.7.0